Extract rten_tensor::tensor::tests into own file
Move the large `tests` submodule into its own file.
